Mp4 - allow signed TimeOffset for samples (when version==1)

+++ src/libse/ContainerFormats/Mp4/TimeSegment.cs
@@ -4,7 +4,7 @@
     {
         public uint? Duration { get; set; }
         public uint? Size { get; set; }
-        public uint? TimeOffset { get; set; }
+        public long? TimeOffset { get; set; }
         public ulong BaseMediaDecodeTime { get; set; }
     }
 }
